Move TrafficDriver from model.commercial into commercial.model.capi.
